requirements.txt
# Core dependencies
streamlit>=1.35.0
pandas>=2.0.0
numpy>=1.26.0
plotly>=5.18.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# LangChain
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.1.0
langgraph>=0.1.0
langgraph-checkpoint>=1.0.0
langgraph-prebuilt>=0.1.0
langsmith>=0.1.0

# OpenAI
openai>=1.20.0
httpx>=0.27.0
tiktoken>=0.5.0

# Neo4j
neo4j>=5.15.0

# Utilities
requests>=2.31.0
tenacity>=8.2.0
PyYAML>=6.0.0
python-dateutil>=2.8.0
pypdf>=4.2.0
python-docx>=1.0.0
